feat: add permission to app
Add user and admin permission to app. Admins can:
- add a contact to a user
- delete any contact
- view all contacts
- update any contacts

+++ app/controllers/contact_controller.py
@@ -1,5 +1,3 @@
-import asyncio
-
 from fastapi import APIRouter, HTTPException, Depends, status, Body, Request
 from sqlalchemy.orm import Session
 from typing import List
@@ -19,44 +17,39 @@ async def create_contact(
     contact_data: CreateContact = Body(...),
     user: User = Depends(get_user),
 ):
+    await check_permission(action="create", resource="contact", user=user)
     try:
-        await check_permission(
-            user_id=str(user.id), action="create", resource="contact"
-        )
-
         return contact_service.create_contact(user.id, contact_data)
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(e))
 
 
 @router.get("/", response_model=List[GetContact])
-def get_all_contacts(
-    db: Session = Depends(get_db),
+async def get_user_contacts(
     user: User = Depends(get_user),
 ):
-    return contact_service.get_all_contacts(user.id, db)
+    return contact_service.get_user_contacts(user.id)
 
 
 @router.get("/{id}", response_model=GetContact)
 def get_contact_by_id(
     id: int,
-    db: Session = Depends(get_db),
     user: User = Depends(get_user),
 ):
     try:
-        return contact_service.get_contact_by_id(id, user.id, db)
+        return contact_service.get_contact_by_id(id, user.id)
     except ValueError as e:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
 
 
 @router.put("/{id}")
-def update_contact(
+async def update_contact(
     id: int,
     contact_data: UpdateContact,
-    db: Session = Depends(get_db),
     user: User = Depends(get_user),
 ):
     try:
+        await check_permission(action="update", resource="contact", user=user)
         contact_service.update_contact(id, user.id, contact_data)
         return {"message": "Contact updated successfully"}
     except ValueError as e:
@@ -64,12 +57,12 @@ def update_contact(
 
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_contact(
+async def delete_contact(
     id: int,
-    db: Session = Depends(get_db),
     user: User = Depends(get_user),
 ):
     try:
+        await check_permission(action="delete", resource="contact", user=user)
         contact_service.delete_contact(id, user.id)
     except ValueError as e:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
